give fuzz it's own matrix
