+++ bson/encoder_test.go
@@ -119,54 +119,6 @@ func TestEncoderEncode(t *testing.T) {
 	})
 }
 
-func Equal(a []byte, b []byte) bool {
-	for i := range a {
-		if a[i] != b[i] {
-			return false
-		}
-	}
-	return true
-}
-
-func FuzzSomething(f *testing.F) {
-	f.Fuzz(func(t *testing.T, a []byte, b []byte) {
-		Equal(a, b)
-	})
-}
-
-func FuzzingSomethingAgain(f *testing.F) {
-	f.Fuzz(func(t *testing.T, a []byte, b []byte) {
-		Equal(a, b)
-	})
-}
-
-func FuzzDecode(f *testing.F) {
-	f.Fuzz(func(t *testing.T, data []byte) {
-		for _, typ := range []func() interface{}{
-			func() interface{} { return new(D) },
-			func() interface{} { return new([]E) },
-			func() interface{} { return new(M) },
-			func() interface{} { return new(interface{}) },
-			func() interface{} { return make(map[string]interface{}) },
-			func() interface{} { return new([]interface{}) },
-		} {
-			i := typ()
-			if err := Unmarshal(data, i); err != nil {
-				return
-			}
-
-			encoded, err := Marshal(i)
-			if err != nil {
-				t.Fatal("failed to marshal", err)
-			}
-
-			if err := Unmarshal(encoded, i); err != nil {
-				t.Fatal("failed to unmarshal", err)
-			}
-		}
-	})
-}
-
 type testMarshaler struct {
 	buf []byte
 	err error
